requirements.txt
# Soulfra Magic Publish System - Python Dependencies
# Install all: pip3 install -r requirements.txt

# Web Framework
flask>=3.0.0
flask-cors>=3.0.0
Werkzeug>=3.0.1

# HTTP Requests
requests>=2.31.0

# HTML/XML Parsing
beautifulsoup4>=4.14.3
lxml>=5.1.0
markdown2>=2.4.0

# AI/ML APIs
openai>=1.55.3
anthropic>=0.40.0
ollama>=0.4.4

# AI Image Generation (Stable Diffusion)
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
invisible_watermark>=0.2.0

# OCR (Text Extraction from Images)
easyocr>=1.7.0

# Image Processing
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3

# Security
cryptography>=41.0.7

# Web Server (Production)
gunicorn>=21.2.0

# Testing (optional)
pytest>=7.4.3
pytest-flask>=1.3.0
